+++ includes/cli/templates/block/index.php
@@ -3,5 +3,10 @@
  * Gutenberg block setup
  */
 
+// enqueue block assets
+wp_register_style( 'aquamin-block-template-slug-style', get_template_directory_uri() . '/dist/blocks/block-library/template-slug/template-slug-style.css', null, '1.0' );
+
 // register the block
-register_block_type_from_metadata( __DIR__ );
+register_block_type( __DIR__, array(
+	'style_handles' => ['aquamin-block-template-slug-style']
+) );