+++ includes/cli/wp-cli.php
@@ -18,18 +18,55 @@ if ( ! defined( 'ABSPATH' ) ) {
 class AQUAMIN_CLI {
 
 	/**
-     * We are asking a question and returning an answer as a string.
+     * Asks a question and return result.
      *
      * @param {string} $question Question to ask the user.
      * @param {string} $guess Optional guess at the user's answer.
+     * @param {string} $default Optional default (i.e. passed as cli arguments)
      * @return string
      */
-    protected function ask( $question, $guess='' ) {
-        fwrite( STDOUT, $question . ' ' );
-		$answer = trim( fgets( STDIN ) );
-		$answer = $answer ? $answer : $guess;
-        return $answer;
+    protected function ask( $question, $guess='', $default=null ) {
+		// if user explicitly passed a boolean as a cli option then default to "y"
+		if ( 'boolean' === gettype( $default ) ) {
+			return 'y';
+		// if user explicitly passed a string as a cli option then use it
+		} elseif ( $default ) {
+			return $default;
+		// if the user hasn't defined something via a cli option
+		} else {
+			// ask the user vie command line
+			fwrite( STDOUT, $question . ' ' );
+			$answer = trim( fgets( STDIN ) );
+			// go with their answer or our guess if they didn't answer
+			$answer = $answer ? $answer : $guess;
+			// send it!
+			return $answer;
+		}
     }
+
+	/**
+	 * Find-and-replace text in files
+	 * 
+	 * @param {array}	$files		File paths.
+	 * @param {array}	$strings	Array like [[replace, find], [replace, find]].
+	 * 
+	 * @return null
+	 */
+	protected function far( $path, $strings ) {
+		$files = new RecursiveIteratorIterator( new RecursiveDirectoryIterator( $path ) );
+		if ( ! empty( $files ) ) {
+			foreach( $files as $file ) {
+				// perform find and replace
+				if ( is_file( $file ) ) {
+					$str = file_get_contents( $file );
+					foreach ( $strings as $key => $value ) {
+						$str = str_replace( $value[ 1 ], $value[ 0 ], $str);
+					}
+					file_put_contents( $file, $str );
+				}
+			}
+		}
+	}
 
 	/**
 	 * Creates a component
@@ -45,7 +82,16 @@ class AQUAMIN_CLI {
 	 * : Set slug.
 	 * 
 	 * [--component_dir] 
-	 * : Set directory.
+	 * : Set directory name.
+	 * 
+	 * [--has_js] 
+	 * : Add front-end JavaScript file (set to "n" to prevent)
+	 * 
+	 * [--has_template_part] 
+	 * : Add a PHP template part (set to "n" to prevent)
+	 * 
+	 * [--has_admin_css] 
+	 * : Add CSS for editor (set to "n" to prevent)
      *
      * ## EXAMPLES
      *
@@ -71,35 +117,36 @@ class AQUAMIN_CLI {
 		
 		$title = 'My Component';
 		$guess = $title;
+
 		$component[ 'component_title' ] = array(
-			$assoc_args[ 'component_title' ] ?? $this->ask( "Title [$guess]:", $guess ),
+			$this->ask( "Title [$guess]:", $guess, $assoc_args[ 'component_title' ] ?? '' ),
 			'template-title',
 		);
 		$title = $component[ 'component_title' ][ 0 ];
 		
 		$guess = sanitize_title( $title );
 		$component[ 'component_slug' ] = array(
-			$assoc_args[ 'component_slug' ] ?? $this->ask( "Slug [$guess]:", $guess ),
+			$this->ask( "Slug [$guess]:", $guess, $assoc_args[ 'component_slug' ] ?? '' ),
 			'template-slug',
 		);
 		$slug = $component[ 'component_slug' ][ 0 ];
 		
 		$guess = $slug;
 		$component[ 'component_dir' ] = array(
-			$assoc_args[ 'component_dir' ] ?? $this->ask( "Directory [$guess]:", $guess ),
+			$this->ask( "Directory [$guess]:", $guess, $assoc_args[ 'component_dir' ] ?? '' ),
 			'_template-component',
 		);
 
 		$exclude = array();
 
-		$ask_admin_css = $this->ask('Has admin CSS? [y/N]');
-		$exclude['admin_css'] = strtolower( $ask_admin_css ?? '' ) === 'y' ? false : 'template-slug-editor.css';
-		
-		$ask_js = $this->ask('Has front-end JavaScript? [y/N]');
-		$exclude['js'] = strtolower( $ask_js ?? '' ) === 'y' ? false : 'template-slug-script.js';
+		$ask_js = $this->ask( 'Has front-end JavaScript? [y/N]', 'n', $assoc_args[ 'has_js' ] );
+		$exclude['js'] = strtolower( $ask_js ) === 'y' ? false : 'template-slug-script.js';
 
-		$ask_template_part = $this->ask('Has get_template_part() PHP? [y/N]');
-		$exclude['template_part'] = strtolower( $ask_template_part ?? '' ) === 'y' ? false : 'template-slug-markup.php';
+		$ask_template_part = $this->ask( 'Has PHP template part? [y/N]', 'n', $assoc_args[ 'has_template_part' ] );
+		$exclude['template_part'] = strtolower( $ask_template_part ) === 'y' ? false : 'template-slug-markup.php';
+
+		$ask_admin_css = $this->ask( 'Has admin CSS? [y/N]', 'n', $assoc_args[ 'has_admin_css' ] );
+		$exclude['admin_css'] = strtolower( $ask_admin_css ) === 'y' ? false : 'template-slug-editor.css';
 
 		/**
 		 * Identify paths/dirs
@@ -176,7 +223,7 @@ class AQUAMIN_CLI {
 	 * : Set slug.
 	 * 
 	 * [--block_dir] 
-	 * : Set directory.
+	 * : Set directory name.
 	 * 
 	 * [--block_namespace] 
 	 * : Set namespace.
@@ -195,8 +242,17 @@ class AQUAMIN_CLI {
 	 * 
 	 * [--inner_block_desc] 
 	 * : Set inner block description.
+	 * 
+	 * [--has_js] 
+	 * : Add front-end JavaScript file (set to "n" to prevent)
+	 * 
+	 * [--is_dynamic] 
+	 * : Make it a dynamic block (set to "n" to prevent).
+	 * 
+     * [--has_inner_block] 
+	 * : Include an inner block (set to "n" to prevent).
      *
-     * ## EXAMPLES
+	 * ## EXAMPLES
      *
      * wp aquamin create block
 	 *
@@ -221,48 +277,48 @@ class AQUAMIN_CLI {
 		$title = 'My Block';
 		$guess = $title;
 		$block[ 'block_title' ] = array(
-			$assoc_args[ 'block_title' ] ?? $this->ask( "Title [$guess]:", $guess ),
+			$this->ask( "Title [$guess]:", $guess, $assoc_args[ 'block_title' ] ?? '' ),
 			'template-title',
 		);
 		$title = $block[ 'block_title' ][ 0 ];
 		
 		$guess = sanitize_title( $title );
 		$block[ 'block_slug' ] = array(
-			$assoc_args[ 'block_slug' ] ?? $this->ask( "Slug [$guess]:", $guess ),
+			$this->ask( "Slug [$guess]:", $guess, $assoc_args[ 'block_slug' ] ?? '' ),
 			'template-slug',
 		);
 		$slug = $block[ 'block_slug' ][ 0 ];
 		
 		$guess = $slug;
 		$block[ 'block_dir' ] = array(
-			$assoc_args[ 'block_dir' ] ?? $this->ask( "Directory [$guess]:", $guess ),
+			$this->ask( "Directory [$guess]:", $guess, $assoc_args[ 'block_dir' ] ?? '' ),
 			'_template-block',
 		);
 		
 		$guess = str_replace(' ', '', $title);
 		$block[ 'block_namespace' ] = array(
-			$assoc_args[ 'block_namespace' ] ?? $this->ask( "Namespace [$guess]:", $guess ),
+			$this->ask( "Namespace [$guess]:", $guess, $assoc_args[ 'block_namespace' ] ?? '' ),
 			'TemplateNamespace',
 		);
 
 		$guess = 'The ' . $title . ' block.';
 		$block[ 'block_desc' ] = array(
-			$assoc_args[ 'block_desc' ] ?? $this->ask( "Description [$guess]:", $guess ),
+			$this->ask( "Description [$guess]:", $guess, $assoc_args[ 'block_desc' ] ?? '' ),
 			'template-desc',
 		);
 
-		$ask_js = $this->ask('Has front-end JavaScript? [y/N]');
-		$has_js = strtolower( $ask_js ?? '' ) === 'y' ? true : false;
+		$ask_js = $this->ask( 'Has front-end JavaScript? [y/N]', 'n', $assoc_args[ 'has_js' ] );
+		$has_js = strtolower( $ask_js ) === 'y' ? true : false;
 
-		$ask_dynamic = $this->ask('Dynamic? [y/N]');
-		$is_dynamic = strtolower( $ask_dynamic ?? '' ) === 'y' ? true : false;
+		$ask_dynamic = $this->ask( 'Dynamic? [y/N]', 'n', $assoc_args[ 'is_dynamic' ] );
+		$is_dynamic = strtolower( $ask_dynamic ) === 'y' ? true : false;
 
 		$has_inner_block = false;
 		$inner_slug = '';
 		if ( ! $is_dynamic ) {
 
-			$ask_inner_block = $this->ask('Add inner block? [y/N]');
-			$has_inner_block = strtolower( $ask_inner_block ?? '' ) === 'y' ? true : false;
+			$ask_inner_block = $this->ask( 'Add inner block? [y/N]', 'n', $assoc_args[ 'has_inner_block' ] );
+			$has_inner_block = strtolower( $ask_inner_block ) === 'y' ? true : false;
 			$inner_block = array();
 			if ( $has_inner_block ) {
 	
@@ -329,24 +385,36 @@ class AQUAMIN_CLI {
 
 		// if we have front-end scripts
 		if ( $has_js ) {
+			// copy the script into the block directory
 			copy( $cli_path . '/common/template-slug-script.js', $block_path . 'template-slug-script.js' );
+			// enqueue it
+			$script_replacements = array(
+				'index.php-1' => array(
+					<<< EOD
+					wp_register_script( 'aquamin-block-template-slug-script', get_template_directory_uri() . '/dist/blocks/block-library/template-slug/template-slug-script.js', null );
+					wp_register_style( 'aquamin-block-template-slug-style', get_template_directory_uri() . '/dist/blocks/block-library/template-slug/template-slug-style.css', null, '1.0' );
+					EOD,
+					<<< EOD
+					wp_register_style( 'aquamin-block-template-slug-style', get_template_directory_uri() . '/dist/blocks/block-library/template-slug/template-slug-style.css', null, '1.0' );
+					EOD
+				),
+				'index.php-2' => array(
+					<<< EOD
+					'view_script_handles' => ['aquamin-block-template-slug-script'],
+						'style_handles' => ['aquamin-block-template-slug-style']
+					EOD,
+					<<< EOD
+					'style_handles' => ['aquamin-block-template-slug-style']
+					EOD
+				),
+			);
+			$this->far( $block_path, $script_replacements );
 		}
 
 		/**
 		 * Find and replace strings
 		 */
-		$block_files = new RecursiveDirectoryIterator( $block_path );
-		// loop through new block's directory
-		foreach( new RecursiveIteratorIterator( $block_files ) as $file ) {
-			// perform find and replace
-			if ( is_file( $file ) ) {
-				$str = file_get_contents( $file );
-				foreach ( $block as $key => $value ) {
-					$str = str_replace( $value[ 1 ], $value[ 0 ], $str);
-				}
-				file_put_contents( $file, $str );
-			}
-		}
+		$this->far( $block_path, $block );
 		
 		/**
 		 * Setup inner blocks
@@ -354,15 +422,9 @@ class AQUAMIN_CLI {
 		if ( $has_inner_block ) {
 
 			// register/setup the inner block within parent files
-			$inner_block_removes = array(
-				array(
-					'path' => $block_path . 'index.js',
-					'replace' =><<< EOD
-					/**
-					 * Import dependencies
-					 */
-					EOD,
-					'with' =><<< EOD
+			$inner_block_replacements = array(
+				'index.js' => array(
+					<<< EOD
 					/**
 					 * Register inner blocks
 					 */
@@ -371,39 +433,35 @@ class AQUAMIN_CLI {
 					/**
 					 * Import dependencies
 					 */
-					EOD
-				),
-				array(
-					'path' => $block_path . 'index.php',
-					'replace' =><<< EOD
-					// register the block
-					register_block_type_from_metadata( __DIR__ );
 					EOD,
-					'with' =><<< EOD
-					// register the block
-					register_block_type_from_metadata( __DIR__ );
-
-					// register inner blocks
-					register_block_type_from_metadata( __DIR__ . '/template-item-slug' );
+					<<< EOD
+					/**
+					 * Import dependencies
+					 */
 					EOD
 				),
-				array(
-					'path' => $block_path . 'template-slug-edit.js',
-					'replace' => '<InnerBlocks />',
-					'with' =><<< EOD
-					<InnerBlocks
-									template={[['aquamin/template-item-slug']]}
-									allowedBlocks={['aquamin/template-item-slug']}
-								/>
+				'index.php' => array(
+					<<< EOD
+					// register inner blocks
+					register_block_type( __DIR__ . '/template-item-slug' );
+					
+					// register the block
+					EOD,
+					<<< EOD
+					// register the block
 					EOD
+				),
+				'template-slug-edit.js' => array(
+					<<< EOD
+					<InnerBlocks
+						template={[['aquamin/template-item-slug']]}
+						allowedBlocks={['aquamin/template-item-slug']}
+					/>
+					EOD,
+					'<InnerBlocks />'
 				),
 			);
-			foreach( $inner_block_removes as $remove ) {
-				$file = $remove[ 'path' ];
-				$str = file_get_contents( $file );
-				$str = str_replace( $remove[ 'replace'], $remove[ 'with'], $str );
-				file_put_contents( $file, $str );	
-			}
+			$this->far( $block_path, $inner_block_replacements );
 
 			// copy the inner blocks directory
 			$template_inner_dir = 'block-inner';
@@ -411,22 +469,13 @@ class AQUAMIN_CLI {
 			$wp_filesystem->mkdir( $block_inner_path );
 			copy_dir( $cli_path . $template_inner_dir, $block_inner_path );
 			
-			// loop through block's directory (including parent block) to find/replace text
-			foreach( new RecursiveIteratorIterator( $block_files ) as $file ) {
-				if ( is_file( $file ) ) {
-					$str = file_get_contents( $file );
-					foreach ( $inner_block as $key => $value ) {
-						$str = str_replace( $value[ 1 ], $value[ 0 ], $str);
-
-					}
-					file_put_contents( $file, $str );
-				}
-			}			
+			// loop through block's directory (will include parent block) to find/replace text
+			$this->far( $block_path, $inner_block );
 
 		}
 
 		// loop through the block's directory and replace file prefixes
-		foreach( new RecursiveIteratorIterator( $block_files ) as $file ) {
+		foreach( new RecursiveIteratorIterator( new RecursiveDirectoryIterator( $block_path ) ) as $file ) {
 			if ( is_file( $file ) ) {
 				$file_name = basename( $file );
 				if ( 'template-slug' === substr( $file_name, 0, 13 ) ) {