+++ docs/changelog.md
@@ -10,6 +10,10 @@ search_exclude: true
 
 # Changelog
 
+## 5.0.0 - May 17 2023
+* Set block assets to appear on front-end only on pages where they're used.
+* Update wp cli to support some boolean options.
+
 ## 4.3.0 - May 12 2023
 * Added custom embed scripts option page under Settings > Custom Scripts.
 * Change wp cli `wp aquamin block` command to be `wp aquamin create block` so it makes more sense.